requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
streamlit>=1.20.0
pytest>=7.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
mysql-connector-python>=8.0.0
redis>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4